requirements.txt
flask
cohere
langchain
langchain-community
langchain-huggingface
sentence-transformers
python-dotenv
